Upload user dir from test runs (#13326)
* Add archive logs step to linux build

* right file

* try

* try this

* correct order

* build artifact

* export

* log

* use tmp

* zip up all

* Remove extra publish

* other scripts

* add test name to dir

+++ scripts/test-extensions-unit.sh
@@ -111,5 +111,7 @@ if [[ "$RUN_DBPROJECT_TESTS" == "true" ]]; then
 	"$INTEGRATION_TEST_ELECTRON_PATH" $LINUX_NO_SANDBOX --extensionDevelopmentPath=$ROOT/extensions/sql-database-projects --extensionTestsPath=$ROOT/extensions/sql-database-projects/out/test --user-data-dir=$VSCODEUSERDATADIR --extensions-dir=$VSCODEEXTDIR --disable-telemetry --disable-crash-reporter --disable-updates --nogpu
 fi
 
-rm -r $VSCODEUSERDATADIR
-rm -r $VSCODEEXTDIR
+if [[ "$NO_CLEANUP" == "" ]]; then
+	rm -r $VSCODEUSERDATADIR
+	rm -r $VSCODEEXTDIR
+fi
